Decouple logic and operator rules
